added custom Exception Handling for pensioner service

+++ PensionManagementPensionerService/Controllers/GuardianController.cs
@@ -32,13 +32,18 @@ namespace PensionManagementPensionerService.Controllers
             {
                 _logger.LogInformation("Attempting to retrieve all guardian details.");
                 var result = await _guardianRepository.GetAllGuardianDetails();
+                if (result.Count() == 0)
+                {
+                    throw new PensionerServiceException("No guardian details found.");
+                }
+                
                 _logger.LogInformation("Successfully retrieved all guardian details");
                 return Ok(_mapper.Map<List<GuardianResponseDTO>>(result));
             }
-            catch (EmptyResultException ex)
+            catch (PensionerServiceException ex)
             {
                 _logger.LogError("Empty result returned while retrieving guardian details");
-                return NotFound(ex.Message);
+                return StatusCode(404, ex.Message);
             }
             catch (Exception ex)
             {
@@ -54,13 +59,18 @@ namespace PensionManagementPensionerService.Controllers
             {
                 _logger.LogInformation("Attempting to retrieve guardian details by guardian Id.");
                 var result = await _guardianRepository.GetGuardianById(guardianId);
+                if(result == null)
+                {
+                    throw new PensionerServiceException("No guardian details found for the given guardianID.");
+                }
                 _logger.LogInformation("Successfully retrieved guardian details by guardian Id: {@result}", result.GuardianId);
                 return Ok(_mapper.Map<GuardianResponseDTO>(result));
             }
-            catch (NotFoundException ex)
+            catch (PensionerServiceException ex)
             {
                 _logger.LogError("No guardian details found.");
-                return NotFound(ex.Message);
+                return StatusCode(404, ex.Message);
+                
             }
             catch (Exception ex)
             {
@@ -75,15 +85,20 @@ namespace PensionManagementPensionerService.Controllers
             try
             {
                 _logger.LogInformation("Attempting to add guardian details.");
+                var existingdetails = _guardianRepository.GetGuadianIdByPensionerId(guardianDetails.PensionerId);
+                if (existingdetails != null)
+                {
+                    throw new PensionerServiceException("A guardian with the same details already exists.");
+                }
                 GuardianDetails request = _mapper.Map<GuardianDetails>(guardianDetails);
                 var result = await _guardianRepository.AddGuardian(request);
                 _logger.LogInformation("Successfully added guardian details : {@result}", result.GuardianId);
                 return Ok(_mapper.Map<GuardianResponseDTO>(result));
             }
-            catch (DuplicateRecordException ex)
+            catch (PensionerServiceException ex)
             {
-                _logger.LogError("Attempted to add a duplicate record");
-                return Conflict(ex.Message);
+                _logger.LogError("A guardian with the same details already exists.");
+                return StatusCode(409, ex.Message);
             }
             catch (Exception ex)
             {
@@ -99,15 +114,21 @@ namespace PensionManagementPensionerService.Controllers
             try
             {
                 _logger.LogInformation("Attempting to update guardian details by guardian Id.");
+                var guardian = await _guardianRepository.GetGuardianById(guardianId);
+                if (guardian == null)
+                {
+                    throw new PensionerServiceException("No guardian details found for the given guardianID.");
+
+                }
                 GuardianDetails request = _mapper.Map<GuardianDetails>(guardianDetails);
                 var result = await _guardianRepository.UpdateGuardianById(guardianId, request);
                 _logger.LogInformation("Successfully updated guardian details by guardian Id {@result}", result.GuardianId);
                  return Ok(_mapper.Map<GuardianResponseDTO>(result));
             }
-            catch (NotFoundException ex)
+            catch (PensionerServiceException ex)
             {
                 _logger.LogError("No guardian details found.");
-                return NotFound(ex.Message);
+                return StatusCode(404, ex.Message);
             }
             catch (Exception ex)
             {
@@ -122,14 +143,19 @@ namespace PensionManagementPensionerService.Controllers
             try
             {
                 _logger.LogInformation("Attempting to delete guardian details by guardian Id.");
+                var result = await _guardianRepository.GetGuardianById(guardianId);
+                if (result == null)
+                {
+                    throw new PensionerServiceException("No guardian details found for the given guardianID.");
+                }
                 _guardianRepository.DeleteGuardianById(guardianId);
                 _logger.LogInformation("Successfully deleted guardian details {@guardianId}", guardianId);
                 return NoContent();
             }
-            catch (NotFoundException ex)
+            catch (PensionerServiceException ex)
             {
                 _logger.LogError("No guardian details found.");
-                return NotFound(ex.Message);
+                return StatusCode(404, ex.Message);
             }
             catch (Exception ex)
             {
@@ -142,24 +168,28 @@ namespace PensionManagementPensionerService.Controllers
         public async Task<ActionResult> GetGuardianIdByPensionerId(Guid pensionerId)
         {
             try
-            {
+            { 
                 _logger.LogInformation("Attempting to retrieve guardianId by pensioner Id.");
                 var result = await _guardianRepository.GetGuadianIdByPensionerId(pensionerId);
+                if (result == null)
+                {
+                    throw new PensionerServiceException("No guardianId found for the given pensionerID.");
+                }
                 _logger.LogInformation("Successfully retrieved guardianId by pensioner Id: {@result}", result);
                 return Ok(result);
             }
-            catch (NotFoundException ex)
+
+            catch (PensionerServiceException ex)
             {
-                _logger.LogError("No guardian details found.");
-                return NotFound(ex.Message);
+                _logger.LogError("No guardianId found for the given pensionerID.");
+                return StatusCode(404, ex.Message);
             }
             catch (Exception ex)
             {
                 _logger.LogError("An unexpected error occurred while processing the request: {@ErrorMessage}", ex.Message);
                 return StatusCode(500, "An unexpected error occurred while processing the request. Please try again later.");
             }
-
-
+            
         }
     }
 }
