+++ PensionManagementPensionerService/Controllers/PensionPlanController.cs
@@ -1,10 +1,13 @@
 ﻿using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Http.HttpResults;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Server.IIS.Core;
 using Microsoft.EntityFrameworkCore.Migrations.Operations;
 using PensionManagementPensionerService.ExceptionalHandling;
 using PensionManagementPensionerService.Models;
+using PensionManagementPensionerService.Models.Repository.Implementation;
 using PensionManagementPensionerService.Models.Repository.Interfaces;
+using static PensionManagementPensionerService.ExceptionalHandling.PensionerServiceException;
 
 namespace PensionManagementPensionerService.Controllers
 {
@@ -28,13 +31,17 @@ namespace PensionManagementPensionerService.Controllers
             {
                 _logger.LogInformation("Attempting to retrieve all pension plan details.");
                 var result = await _pensionPlanRepository.GetAllPensionPlans();
+                if (result.Count() == 0)
+                {
+                    throw new PensionerServiceException("No pensionplan details found.");
+                }
                 _logger.LogInformation("Successfully retrieved all pensionplan details");
                 return Ok(result);
             }
-            catch (EmptyResultException ex)
+            catch (PensionerServiceException ex)
             {
                 _logger.LogError("Empty result returned while retrieving pensionplan details");
-                return NotFound(ex.Message);
+                return StatusCode(404, ex.Message);
             }
             catch (Exception ex)
             {
@@ -42,6 +49,7 @@ namespace PensionManagementPensionerService.Controllers
                 return StatusCode(500, "An unexpected error occurred while processing the request. Please try again later.");
             }
         }
+
 
         [HttpGet("{pensionPlanId}")]
         public async Task<IActionResult> GetPensionPlanById(Guid pensionPlanId)
@@ -50,13 +58,17 @@ namespace PensionManagementPensionerService.Controllers
             {
                 _logger.LogInformation("Attempting to retrieve pension plan details by pension plan Id.");
                 var result = await _pensionPlanRepository.GetPensionPlanById(pensionPlanId);
+                if (result == null)
+                {
+                    throw new PensionerServiceException("No pension plan details found for the given pensionPlanID.");
+                }
                 _logger.LogInformation("Successfully retrieved pension plan details by Pensionplan Id : {@result}", result.PensionPlanId);
                 return Ok(result);
             }
-            catch (NotFoundException ex)
+            catch (PensionerServiceException ex)
             {
                 _logger.LogError("No pension plan details found.");
-                return NotFound(ex.Message);
+                return StatusCode(404, ex.Message);
             }
             catch (Exception ex)
             {
@@ -64,6 +76,7 @@ namespace PensionManagementPensionerService.Controllers
                 return StatusCode(500, "An unexpected error occurred while processing the request. Please try again later.");
             }
         }
+        
 
         [HttpPost]
         public async Task<ActionResult<PensionPlanDetails>> AddPensionPlan([FromBody] PensionPlanDetails pensionPlanDetails)
@@ -71,20 +84,27 @@ namespace PensionManagementPensionerService.Controllers
             try
             {
                 _logger.LogInformation("Attempting to add all pension plan details.");
+                var existingdetails = _pensionPlanRepository.GetPensionPlanById(pensionPlanDetails.PensionPlanId);
+                if (existingdetails != null)
+                {
+                    throw new PensionerServiceException("A pension Plan with the same details already exists.");
+                }
                 var result = await _pensionPlanRepository.AddPensionPlan(pensionPlanDetails);
                 _logger.LogInformation("Successfully added pension plan details {@result}", result.PensionPlanId);
                 return Ok(result);
 
             }
-            catch (DuplicateRecordException ex)
+            catch (PensionerServiceException ex)
             {
                 _logger.LogError("Attempted to add a duplicate record");
-                return Conflict(ex.Message);
+                return StatusCode(409, ex.Message);
             }
             catch (Exception ex)
+
             {
                 _logger.LogError("An unexpected error occurred while processing the request: {@ErrorMessage}", ex.Message);
                 return StatusCode(500, "An unexpected error occurred while processing the request. Please try again later.");
+                
             }
 
         }
@@ -95,19 +115,27 @@ namespace PensionManagementPensionerService.Controllers
             try
             {
                 _logger.LogInformation("Attempting to update pension plan details by pensionplan id.");
+                var pensionPlan = await _pensionPlanRepository.GetPensionPlanById(pensionPlanId);
+                if (pensionPlan == null)
+                {
+                    throw new PensionerServiceException("No pensionPlan details found for the given pensionPlanID.");
+
+                }
                 var result = await _pensionPlanRepository.UpdatePensionPlanById(pensionPlanId,pensionPlanDetails);
                 _logger.LogInformation("Successfully updated pensionplan details by Pensionplan Id: {@result}", result.PensionPlanId);
                 return Ok(result);
             }
-            catch (NotFoundException ex)
+            catch (PensionerServiceException ex)
             {
                 _logger.LogError("No pension plan details found.");
-                return NotFound(ex.Message);
+                return StatusCode(404, ex.Message);
             }
             catch (Exception ex)
+            
             {
                 _logger.LogError("An unexpected error occurred while processing the request: {@ErrorMessage}", ex.Message);
                 return StatusCode(500, "An unexpected error occurred while processing the request. Please try again later.");
+                
             }
         }
 
@@ -117,19 +145,26 @@ namespace PensionManagementPensionerService.Controllers
             try
             {
                 _logger.LogInformation("Attempting to delete pensionplan details by pensionplan id.");
+                var result = await _pensionPlanRepository.GetPensionPlanById(pensionPlanId);
+                if(result == null)
+                {
+                    throw new PensionerServiceException("No pension plan details found for the given pensionplanID.");
+                }
                 _pensionPlanRepository.DeletePensionPlanById(pensionPlanId);
                 _logger.LogInformation("Successfully deleted pensioneplan details by Pensionplan Id {pensionPlanId}.", pensionPlanId );
                 return NoContent();
              }
-            catch (NotFoundException ex)
+            catch (PensionerServiceException ex)
             {
                 _logger.LogError("No pension plan details found.");
-                return NotFound(ex.Message);
+                return StatusCode(404, ex.Message);
             }
             catch (Exception ex)
+            
             {
                 _logger.LogError("An unexpected error occurred while processing the request: {@ErrorMessage}", ex.Message);
                 return StatusCode(500, "An unexpected error occurred while processing the request. Please try again later.");
+                
             }
         }
