+++ PensionManagementPensionerService/Controllers/PensionerController.cs
@@ -4,6 +4,7 @@ using PensionManagementPensionerService.Models.Repository.Interfaces;
 using PensionManagementPensionerService.Models;
 using PensionManagementPensionerService.DTO;
 using PensionManagementPensionerService.ExceptionalHandling;
+using PensionManagementPensionerService.Models.Repository.Implementation;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.AspNetCore.Http.HttpResults;
 using AutoMapper;
@@ -32,13 +33,17 @@ namespace PensionManagementPensionerService.Controllers
             {
                 _logger.LogInformation("Attempting to retrieve all pensioner details.");
                 var result = await _pensionerRepository.GetAllPensionerDetails();
+                if (result.Count() == 0)
+                {
+                    throw new PensionerServiceException("No pensioner details found.");
+                }
                 _logger.LogInformation("Successfully retrieved all pensioner details");             
                 return Ok(_mapper.Map<List<PensionResponseDTO>>(result));
             }
-            catch (EmptyResultException ex)
+            catch (PensionerServiceException ex)
             {
                 _logger.LogError("Empty result returned while retrieving pensioner details");
-                return NotFound(ex.Message);
+                return StatusCode(404, ex.Message);
             }
             catch (Exception ex)
             {
@@ -54,16 +59,21 @@ namespace PensionManagementPensionerService.Controllers
             {
                 _logger.LogInformation("Attempting to retrieve pensioner details by pensioner Id.");
                 var result = await _pensionerRepository.GetPensionerDetailsById(pensionerId);
+                if (result == null)
+                {
+                    throw new PensionerServiceException("No pensioner details found for the given pensionerID.");
+                }
                 _logger.LogInformation("Successfully retrieved pensioner details by Pensioner Id: {@result}", result.PensionerId);
                 return Ok(_mapper.Map<PensionResponseDTO>(result));
             }
-            catch (NotFoundException ex)
+            catch (PensionerServiceException ex)
             {
                 _logger.LogError("No pensioner details found.");
-                return NotFound(ex.Message);
+                return StatusCode(404, ex.Message);
             }
             catch (Exception ex)
             {
+                
                 _logger.LogError("An unexpected error occurred while processing the request: {@ErrorMessage}", ex.Message);
                 return StatusCode(500, "An unexpected error occurred while processing the request. Please try again later.");
             }
@@ -75,16 +85,22 @@ namespace PensionManagementPensionerService.Controllers
             {
                 _logger.LogInformation("Attempting to retrieve pensioner Id by userId.");
                 var result = await _pensionerRepository.GetPensionerIdById(userId);
+                if (result == null)
+                {
+                    throw new PensionerServiceException("No pensioner id found for the given userID.");
+                }
                 _logger.LogInformation("Successfully retrieved pensioner Id by userId : {@result}", result);
                 return Ok(result);
             }
-            catch (NotFoundException ex)
+            catch (PensionerServiceException ex)
+
             {
                 _logger.LogError("No pensioner details found.");
-                return NotFound(ex.Message);
+                return StatusCode(404, ex.Message);
             }
             catch (Exception ex)
             {
+                
                 _logger.LogError("An unexpected error occurred while processing the request: {@ErrorMessage}", ex.Message);
                 return StatusCode(500, "An unexpected error occurred while processing the request. Please try again later.");
             }
@@ -96,16 +112,22 @@ namespace PensionManagementPensionerService.Controllers
             try
             {
                 _logger.LogInformation("Attempting to add pensioner details.");
+                var existingdetails = _pensionerRepository.GetPensionerIdById(pensionerDetails.Id);
+                if (existingdetails != null)
+                {
+                    throw new PensionerServiceException("A pensioner with the same details already exists.");
+                }
                 PensionerDetails request = _mapper.Map<PensionerDetails>(pensionerDetails);
                 var result = await _pensionerRepository.AddPensionerDetails(request);
                 _logger.LogInformation("Successfully added pensioner details. {@result}", result.PensionerId);
                 return Ok(_mapper.Map<PensionResponseDTO>(result));
 
             }
-            catch(DuplicateRecordException ex)
+            catch (PensionerServiceException ex)
             {
-                _logger.LogError("Attempted to add a duplicate record");
-                return Conflict(ex.Message);
+                _logger.LogError("A pensioner with the same details already exists.");
+                return StatusCode(409, ex.Message);
+               
             }
             catch (Exception ex)
             {
@@ -121,19 +143,26 @@ namespace PensionManagementPensionerService.Controllers
             try
             {
                 _logger.LogInformation("Attempting to update pensioner details by pensioner Id.");
+                var pensioner = await _pensionerRepository.GetPensionerDetailsById(pensionerId);
+                if (pensioner == null)
+                {
+                    throw new PensionerServiceException("No pensioner details found for the given pensionerID.");
+
+                }
                 PensionerDetails request = _mapper.Map<PensionerDetails>(pensionerDetails);
                 var result = await _pensionerRepository.UpdatePensionerDetailsById(pensionerId, request);
                 _logger.LogInformation("Successfully updated pensioner details by Pensioner Id {@result}", result.PensionerId);
                 return Ok(_mapper.Map<PensionResponseDTO>(result));
 
             }
-            catch (NotFoundException ex)
+            catch (PensionerServiceException ex)
             {
                 _logger.LogError("No pensioner details found.");
-                return NotFound(ex.Message);
+                return StatusCode(404, ex.Message);
             }
             catch (Exception ex)
             {
+                
                 _logger.LogError("An unexpected error occurred while processing the request: {@ErrorMessage}", ex.Message);
                 return StatusCode(500, "An unexpected error occurred while processing the request. Please try again later.");
             }
@@ -145,17 +174,23 @@ namespace PensionManagementPensionerService.Controllers
             try
             {
                 _logger.LogInformation("Attempting to delete pensioner details by pensioner Id.");
+                var result = await _pensionerRepository.GetPensionerDetailsById(pensionerId);
+                if (result == null)
+                {
+                    throw new PensionerServiceException("No pensioner details found for the given pensionerID.");
+                }
                 _pensionerRepository.DeletePensionerDetailsById(pensionerId);
                 _logger.LogInformation("Successfully deleted pensioner details by Pensioner Id {@pensionerId}.", pensionerId);
                 return NoContent();
             }
-            catch(NotFoundException ex)
+            catch (PensionerServiceException ex)
             {
                 _logger.LogError("No pensioner details found.");
-                return NotFound(ex.Message);
+                return StatusCode(404, ex.Message);
             }
             catch (Exception ex)
             {
+                
                 _logger.LogError("An unexpected error occurred while processing the request: {ErrorMessage}", ex.Message);
                 return StatusCode(500, "An unexpected error occurred while processing the request. Please try again later.");
             }
